✨ feat: Add and refactor components for chat input feature
The code changes include renaming, adding, and deleting files related to the chat input feature. The "Footer" component is removed, and new components like "SaveTopic" and "Footer" are added. Additionally, a new component for mobile layout is included. The SessionList component's index.tsx file is modified to replace an event to prevent double-click on mobile devices.

The changes are done to improve the chat input feature by adding necessary components, removing unnecessary ones, and enhancing the mobile experience by preventing double-clicks.
