+++ src/pages/chat/features/Conversation/Input/Footer/index.tsx
@@ -1,0 +1,35 @@
+import { Icon } from '@lobehub/ui';
+import { useTheme } from 'antd-style';
+import { ArrowBigUp, CornerDownLeft } from 'lucide-react';
+import { memo } from 'react';
+import { useTranslation } from 'react-i18next';
+import { Flexbox } from 'react-layout-kit';
+
+import SaveTopic from './SaveTopic';
+
+const Footer = memo(() => {
+  const theme = useTheme();
+  const { t } = useTranslation('common');
+
+  return (
+    <>
+      <Flexbox
+        gap={4}
+        horizontal
+        style={{ color: theme.colorTextDescription, fontSize: 12, marginRight: 12 }}
+      >
+        <Icon icon={CornerDownLeft} />
+        <span>{t('send')}</span>
+        <span>/</span>
+        <Flexbox horizontal>
+          <Icon icon={ArrowBigUp} />
+          <Icon icon={CornerDownLeft} />
+        </Flexbox>
+        <span>{t('warp')}</span>
+      </Flexbox>
+      <SaveTopic />
+    </>
+  );
+});
+
+export default Footer;
