+++ src/pages/chat/features/Conversation/Input/Mobile.tsx
@@ -1,0 +1,12 @@
+import { ReactNode, memo } from 'react';
+
+interface ChatInputMobileLayoutProps {
+  children: ReactNode;
+  expand?: boolean;
+}
+
+const ChatInputMobileLayout = memo<ChatInputMobileLayoutProps>(({ children }) => {
+  return <div>{children}</div>;
+});
+
+export default ChatInputMobileLayout;
